moved Test code examples
