--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MRColorAction.java
@@ -1,106 +1,0 @@
-/* Copyright (c) 2015 Qualcomm Technologies Inc
-
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted (subject to the limitations in the disclaimer below) provided that
-the following conditions are met:
-
-Redistributions of source code must retain the above copyright notice, this list
-of conditions and the following disclaimer.
-
-Redistributions in binary form must reproduce the above copyright notice, this
-list of conditions and the following disclaimer in the documentation and/or
-other materials provided with the distribution.
-
-Neither the name of Qualcomm Technologies Inc nor the names of its contributors
-may be used to endorse or promote products derived from this software without
-specific prior written permission.
-
-NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS
-LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
-THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
-FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
-
-package org.firstinspires.ftc.teamcode;
-
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.ColorSensor;
-import com.qualcomm.robotcore.hardware.Servo;
-
-/*
- *
- * This is an example LinearOpMode that shows how to use
- * a Modern Robotics Color Sensor.
- *
- * The op mode assumes that the color sensor
- * is configured with a name of "sensor_color".
- *
- * Example of setting action if color sensor reads target value
- *
- * Use Android Studio to Copy this Class, and Paste it into your team's code folder with a new name.
- * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
- */
-@TeleOp(name = "MR ColorAction", group = "Test")
-//@Disabled
-public class MRColorAction extends LinearOpMode {
-
-    // Initialize Hardware Device Objects
-  ColorSensor colorSensor;
-  Servo servoHandL;
-
-    //Create and set default servo positions variables.
-    //Possible servo values: 0.0 - 1.0  For CRServo 0.5=stop greater or less than will spin in that direction
-    double CLOSED = 0.1;
-    double OPEN = 1.0;
-    double NEUTRAL = 0.5;
-
-    // Declare variables
-    double blueTARGET = 20; //determine best value for reading blue sensor true
-    double redTARGET = 20;
-
-    @Override
-  public void runOpMode() throws InterruptedException {
-
-    // Map devices to robot hardware
-        colorSensor = hardwareMap.colorSensor.get("sensor_color");
-        servoHandL = hardwareMap.servo.get("servoHandL");
-
-
-    // wait for the start button to be pressed.
-    waitForStart();
-
-    // while the op mode is active, loop and read the RGB data.
-    // Note we use opModeIsActive() as our loop condition because it is an interruptible method.
-    while (opModeIsActive()) {
-
-      // check for blue present greater than Target value
-        if (colorSensor.blue()>blueTARGET){
-            //do this
-            servoHandL.setPosition(OPEN);
-      // check for red present greater than Target value
-        }else if (colorSensor.red()>redTARGET) {
-            //do this
-            servoHandL.setPosition(CLOSED);
-        }else { //no color sensor reads target value
-            //do this
-            servoHandL.setPosition(NEUTRAL);
-        }
-
-      // send the info back to driver station using telemetry function.
-      telemetry.addData("Red  ", colorSensor.red());
-      telemetry.addData("Blue ", colorSensor.blue());
-
-      telemetry.update();
-    }
-  }
-}
